responsive max range
