+++ client/Assets/Scripts/UI/LeftMMTouchRepositionableJoystick.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: f26d8b5a4a0ac49a6b4edc2ac2d8b060
+guid: 4a3319a63c2104ad6bfb6ddb4b8b219b
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
